Address review comments and clean up tests after rebase - move createPlatformFiles into the versions reducer
