+++ src/core/installAddon.js
@@ -226,7 +226,7 @@ type WithInstallHelpersProps = {|
 
 type WithInstallHelpersInternalProps = {|
   ...WithInstallHelpersProps,
-  currentVersion: AddonVersionType,
+  currentVersion: AddonVersionType | null,
   dispatch: DispatchFunc,
   location: ReactRouterLocationType,
 |};
@@ -305,7 +305,7 @@ export class WithInstallHelpers extends React.Component<WithInstallHelpersIntern
       return Promise.resolve();
     }
 
-    if (!addon) {
+    if (!addon || !currentVersion) {
       return Promise.resolve();
     }
 
@@ -403,6 +403,8 @@ export class WithInstallHelpers extends React.Component<WithInstallHelpersIntern
       location,
       userAgentInfo,
     } = this.props;
+
+    invariant(currentVersion, 'currentVersion is required');
 
     const { guid, name, type } = addon;
     const { platformFiles } = currentVersion;
