+++ tests/unit/amo/components/TestPermissionUtils.js
@@ -1,7 +1,7 @@
 import UAParser from 'ua-parser-js';
 
 import { PermissionUtils } from 'amo/components/PermissionsCard/permissions';
-import { createPlatformFiles } from 'core/reducers/addons';
+import { createPlatformFiles } from 'core/reducers/versions';
 import {
   fakeI18n,
   fakeVersion,
